Checking that records aren't returned with different names
This is because my Airtable search is done by prefixes. I am going this route, rather than trying to make the Airtable query an exact match, as I saw a record in Airtable that was tidied up compared to Trello; the Trello card name was missing the end parentheses while the Airtable record included it. My method of prefix searching will allow that.